sbf: move loader to subpackage
