+++ pkg/sbf/loader/loader_test.go
@@ -1,4 +1,4 @@
-package sbf
+package loader
 
 import (
 	"debug/elf"
