+++ pkg/sbf/loader/tests/empty_rodata.so
