+++ pkg/sbf/loader/tests/pass_stack_reference.so
