+++ pkg/sbf/loader/tests/reloc_64_64.so
