+++ pkg/sbf/loader/tests/reloc_64_relative.so
