+++ pkg/sbf/loader/tests/rodata.so
