+++ pkg/sbf/loader/tests/rodata_high_vaddr.so
