+++ pkg/sbf/loader/tests/scratch_registers.so
