+++ pkg/sbf/loader/tests/syscall_static.so
